move types to type folder
