+++ src/utils/asset.ts
@@ -1,19 +1,4 @@
+import { Asset } from '../types/asset';
 import testnet_assets from '../../config/testnet/assets.json';
-
-export interface Asset {
-    id: string;
-    symbol: string;
-    name: string;
-    is_stablecoin: boolean;
-    contracts: Contract[];
-}
-
-export interface Contract {
-    contract_address?: string;
-    chain_id?: number;
-    contract_decimals?: number;
-    symbol?: string;
-    decimals?: number;
-}
 
 export const getAssets: () => Asset[] = () => testnet_assets;
